Convert from a list of tree rings polygons (L) to data frame to a data frame with no.,year,x,y,r(radius),radian(center angle),degree. The data frame is sorted by degree(0 to 360).

Convert from a list of tree rings polygons (L) to data frame to a data frame with no.,year,x,y,r(radius),radian(center angle),degree. The data frame is sorted by degree(0 to 360).

Llist2dataframe(L)

Arguments

L

list of tree ring lines

Value

data frame

Examples

L_ <- Llist2dataframe(TR@L)
head(L_) ; tail(L_)
#>    i yr         x          y        r       rad       deg
#> 66 1  0 -7619.734  -50.31881 7619.900 -3.134989 -179.6216
#> 65 1  0 -7614.858 -126.71623 7615.912 -3.124954 -179.0466
#> 64 1  0 -7611.607 -187.67161 7613.920 -3.116942 -178.5876
#> 63 1  0 -7606.730 -280.32379 7611.894 -3.104757 -177.8895
#> 62 1  0 -7601.041 -367.28680 7609.910 -3.093310 -177.2336
#> 61 1  0 -7598.603 -433.11862 7610.937 -3.084654 -176.7377
#>        i  yr         x         y         r      rad       deg
#> 8988  38 185  46.38425  81.44112  93.72382 1.053073  60.33661
#> 71022 38 185  12.88853 100.37348 101.19758 1.443089  82.68294
#> 61022 38 185 -20.60720  96.00448  98.19122 1.782236 102.11462
#> 51023 38 185 -52.64658  85.81013 100.67294 2.121101 121.53012
#> 41024 38 185 -75.94795  52.31441  92.22195 2.538410 145.44018
#> 31026 38 185 -91.96765  17.36235  93.59219 2.955001 169.30909