A sample object of class TreeRingShape, shapefile paths and column names only.

The full data set contains tree ring shape data for Abies_277_h400 sampled from Tateyama, central Japan. Its disk image and shape files can be download from https://www.sanchikanri.com/treering/Abies_277_h400.zip

TR_

Format

An object of class classTreeRingShape of length 1.

Examples

# Access basic information about the TreeRingShape object
TR_<-new_classTreeRingShape(
P_filename='Abies_277_h400_TreeRing_Points.shp',
L_filename='Abies_277_h400_TreeRing_Representative.shp',
L2_filename='Abies_277_h400_TreeRing.shp',
P_id.tag='id',
P_ring.tag='ring',
L_ring.tag='ring')
slotNames(TR_)
#>  [1] "P_filename"  "P_id.tag"    "P_ring.tag"  "P"           "P00"        
#>  [6] "n_id"        "YR_P"        "L_filename"  "L_ring.tag"  "L"          
#> [11] "L_"          "YR_L"        "ln"          "L2_filename" "L2"         
#> [16] "n_YR"       
str(TR_)
#> Formal class 'classTreeRingShape' [package "TreeRingShape"] with 16 slots
#>   ..@ P_filename : chr "Abies_277_h400_TreeRing_Points.shp"
#>   ..@ P_id.tag   : chr "id"
#>   ..@ P_ring.tag : chr "ring"
#>   ..@ P          :'data.frame':	0 obs. of  0 variables
#> Formal class 'data.frame' [package "methods"] with 4 slots
#>   .. .. ..@ .Data    : list()
#>   .. .. ..@ names    : chr(0) 
#>   .. .. ..@ row.names: int(0) 
#>   .. .. ..@ .S3Class : chr "data.frame"
#>   ..@ P00        : num(0) 
#>   ..@ n_id       : num(0) 
#>   ..@ YR_P       : num(0) 
#>   ..@ L_filename : chr "Abies_277_h400_TreeRing_Representative.shp"
#>   ..@ L_ring.tag : chr "ring"
#>   ..@ L          : list()
#>   ..@ L_         :'data.frame':	0 obs. of  0 variables
#> Formal class 'data.frame' [package "methods"] with 4 slots
#>   .. .. ..@ .Data    : list()
#>   .. .. ..@ names    : chr(0) 
#>   .. .. ..@ row.names: int(0) 
#>   .. .. ..@ .S3Class : chr "data.frame"
#>   ..@ YR_L       : num(0) 
#>   ..@ ln         : num(0) 
#>   ..@ L2_filename: chr "Abies_277_h400_TreeRing.shp"
#>   ..@ L2         : list()
#>   ..@ n_YR       : num(0)