Return a vector of sequence of angles between start and end angle 0 to pi -pi to 0

Return a vector of sequence of angles between start and end angle 0 to pi -pi to 0

Usage

# S3 method for deg
seq(deg1, deg2, deg.by = 1)

Arguments

deg1

start angle

deg2

end angle

deg.by

step of sequence

Value

vector of sequence of angles between start and end angle

Examples

seq.deg(170,-170,.5)
#>  [1]  170.0  170.5  171.0  171.5  172.0  172.5  173.0  173.5  174.0  174.5
#> [11]  175.0  175.5  176.0  176.5  177.0  177.5  178.0  178.5  179.0  179.5
#> [21]  180.0 -180.0 -179.5 -179.0 -178.5 -178.0 -177.5 -177.0 -176.5 -176.0
#> [31] -175.5 -175.0 -174.5 -174.0 -173.5 -173.0 -172.5 -172.0 -171.5 -171.0
#> [41] -170.5 -170.0