Sort x,y coordinates of tree ring lines with center angle of each point apply Lsort to list of tree ring lines

Sort x,y coordinates of tree ring lines with center angle of each point apply Lsort to list of tree ring lines

Lsort_all(L)

Arguments

L

a list of tree ring lines (x,y)

Value

a list of tree ring lines (x,y) ordered with center angle of each point

Examples

str(Lsort_all(TR@L))
#> List of 38
#>  $ 0  : num [1:1005, 1:2] -7620 -7615 -7612 -7607 -7601 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 1  : num [1:1830, 1:2] -7602 -7599 -7595 -7588 -7590 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 10 : num [1:1908, 1:2] -7431 -7426 -7422 -7416 -7410 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 23 : num [1:970, 1:2] -7248 -7239 -7231 -7225 -7218 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 28 : num [1:1189, 1:2] -7150 -7150 -7143 -7133 -7132 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 33 : num [1:863, 1:2] -7042 -7035 -7033 -7023 -7016 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 38 : num [1:885, 1:2] -6921 -6923 -6914 -6899 -6897 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 43 : num [1:1353, 1:2] -6810 -6805 -6800 -6791 -6783 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 48 : num [1:1034, 1:2] -6707 -6703 -6701 -6703 -6703 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 53 : num [1:877, 1:2] -6613 -6609 -6607 -6600 -6586 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 58 : num [1:513, 1:2] -6463 -6459 -6451 -6445 -6442 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 63 : num [1:339, 1:2] -6184 -6165 -6147 -6126 -6126 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 68 : num [1:868, 1:2] -5942 -5940 -5926 -5918 -5910 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 73 : num [1:736, 1:2] -5767 -5753 -5748 -5743 -5738 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 78 : num [1:509, 1:2] -5589 -5580 -5572 -5559 -5550 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 83 : num [1:400, 1:2] -5379 -5366 -5362 -5339 -5319 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 88 : num [1:190, 1:2] -5122 -5100 -5061 -5024 -4989 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 93 : num [1:222, 1:2] -4924 -4891 -4844 -4813 -4772 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 98 : num [1:164, 1:2] -4694 -4653 -4604 -4571 -4560 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 103: num [1:157, 1:2] -4397 -4371 -4333 -4283 -4263 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 108: num [1:168, 1:2] -4132 -4100 -4062 -4018 -3995 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 113: num [1:157, 1:2] -3884 -3861 -3841 -3771 -3742 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 118: num [1:156, 1:2] -3561 -3535 -3468 -3436 -3415 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 123: num [1:155, 1:2] -3287 -3270 -3238 -3194 -3153 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 128: num [1:140, 1:2] -3022 -3010 -2967 -2923 -2865 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 133: num [1:135, 1:2] -2777 -2757 -2690 -2620 -2565 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 138: num [1:126, 1:2] -2553 -2539 -2504 -2460 -2405 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 143: num [1:113, 1:2] -2314 -2303 -2282 -2239 -2207 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 148: num [1:99, 1:2] -2046 -2029 -2003 -1979 -1939 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 153: num [1:97, 1:2] -1799 -1784 -1746 -1714 -1697 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 158: num [1:88, 1:2] -1542 -1528 -1496 -1435 -1408 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 163: num [1:84, 1:2] -1292 -1280 -1234 -1205 -1190 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 168: num [1:74, 1:2] -1074 -1059 -1047 -1030 -1012 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 173: num [1:59, 1:2] -849 -849 -843 -811 -774 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 178: num [1:64, 1:2] -637 -642 -634 -608 -574 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 183: num [1:39, 1:2] -324 -313 -287 -255 -249 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 184: num [1:24, 1:2] -156 -149 -133 -128 -124 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"
#>  $ 185: num [1:20, 1:2] -86.1 -71.6 -71.6 -49.7 -30.8 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:2] "X" "Y"